feat: use abbreviated syntax for all imports
The motivation to use the abbreviated syntax for all imports
(see [^1], [^2], [^3], and [^4]) is that on the one hand it's
more consistent to read the disassembly that way, as for example
all functions will be printed as `(func $name ...)` no matter if
they are imported or declared in here, and same for the globals,
memories, and tables. But there's a second reason to prefer this
format, because it allows us to use the inline syntax for exports
and therefore avoid having to print that information separately,
but instead it'll be right where it's relevant.

[^1]: https://webassembly.github.io/spec/core/text/modules.html#text-func-abbrev
[^2]: https://webassembly.github.io/spec/core/text/modules.html#text-global-abbrev
[^3]: https://webassembly.github.io/spec/core/text/modules.html#text-mem-abbrev
[^4]: https://webassembly.github.io/spec/core/text/modules.html#text-table-abbrev

Refs: #56