feat: add support for inline export notation
This adds a new `IExportMetadata` interface, which provides export
names for functions, globals, memories, and tables, and support for
it in the `DevToolsNameGenerator`. When such an object is installed
on the `WasmDisassembler#exportMetadata` property, the disassembly
will not print exports explicitly via `(export "name" ...)`, but
instead put the export names inline with the functions, globals,
memories and tables. This saves space in the disassembly and puts
information about exports close to the relevant places.

Refs: #56