feat: support some extended name subsections
This adds initial support for extended "name" subsections, including
type, table, memory, and global names. While the proposal is currently
at Stage 1 only, it's already being implemented in Binaryen and thus
in Emscripten as well as wabt, so it makes sense to support this early
on.

As a drive-by-fix, this also removes some of the code duplication that
we had in the `NameSectionNameResolver` and the `DevToolsNameResolver`
classes.

Bug: https://crbug.com/1137335
Proposal: https://github.com/WebAssembly/extended-name-section